requirements.txt
flask==2.3.3
twilio==8.5.0
openai==1.3.0
python-dotenv==1.0.0
ngrok==0.12.0 